api/requirements.txt
adal==1.2.2
amqp==2.5.0
asn1crypto==0.24.0
awscli==1.16.257
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.23
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.5
azure-datalake-store==0.0.46
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.5
boto3==1.9.247
botocore==1.12.247
bs4==0.0.1
cachetools==3.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
dnspython==1.16.0
docopt==0.4.0
docutils==0.15.2
Flask==1.0.2
flask-api==1.1
Flask-Cors==3.0.4
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-language==1.3.0
google-cloud-speech==1.2.0
google-cloud-storage==1.19.0
google-cloud-vision==0.38.0
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.6
imageio==2.6.1
imageio-ffmpeg==0.3.0
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.4
kombu==4.6.3
lxml==4.2.1
mandrill-really-maintained==1.2.4
MarkupSafe==1.0
moviepy==1.0.1
msrest==0.6.8
msrestazure==0.6.1
numpy==1.15.1
oauthlib==3.0.2
opencv-python==4.1.1.26
pandas==0.23.4
pdf2image==1.6.0
Pillow==6.1.0
proglog==0.1.9
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.18
PyJWT==1.7.1
flask-pymongo==2.3.0
pymongo[srv]
PyPDF2==1.26.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==5.1.2
redis==2.10.6
requests==2.20.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
# selenium==3.12.0
six==1.11.0
tqdm==4.36.1
Unidecode==1.0.22
urllib3==1.22
vine==1.3.0
Werkzeug==0.14.1
gunicorn
